backend/requirements.txt
fastapi
uvicorn
exa_py
python-dotenv
